+++ builder/hspkg-builder.nix
@@ -118,7 +118,7 @@ in rec {
   checks = pkgs.recurseIntoAttrs (builtins.mapAttrs
     (_: d: haskellLib.check d)
       (lib.filterAttrs (_: d: d.config.doCheck) components.tests));
-  inherit (package) identifier detailLevel isLocal;
+  inherit (package) identifier detailLevel isLocal isProject;
   inherit setup cabalFile;
   isHaskell = true;
   inherit src;